feat/cart-goods/create: #99 장바구니 조회 기능 구현중

+++ back-end/with-the-developer/src/main/java/com/developer/cartGoods/query/dto/CartGoodsQueryDTO.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Data
 public class CartGoodsQueryDTO {
-    private int goods_code;
+    private long goods_code;
     private int goods_amount;
 //    private List<Image> images;
 }
